pyproject.toml
[project]
name = "daseen"
description = "Daseen"
dynamic = ["version"]
authors = [
    { name = "Kai Buschulte", email = "kb@beebucket.de" }
]
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.36.13",
    "django>=5.0.6",
    "django-cors-headers>=4.3.1",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "drf-nested-routers>=0.94.1",
    "drf-spectacular>=0.28.0",
    "elasticsearch>=8.17.1",
    "extended-dataset-profile", # version specified in tool.uv.sources
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.4",
    "pymemcache>=4.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "boto3-stubs[s3]>=1.37.0",
    "moto[s3]>=5.1.0",
    "pre-commit>=4.2.0",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.9.0",
    "ruff>=0.9.3",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0+dirty"

[tool.hatch.build.hooks.vcs]
version-file = "backend/daseen/_version.py"

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
daseen = "backend.manage:main"

[tool.ruff]
line-length = 120
extend-exclude = [
    "backend/**/migrations",
    "backend/daseen/_version.py",
]

[tool.ruff.lint]
extend-select = [
    "I", # isort replacement
    "S", # bandit
]

[tool.ruff.lint.per-file-ignores]
"backend/**/tests.py" = ["S101", "S106"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
DJANGO_SETTINGS_MODULE = "daseen.test_settings"
testpaths = ["backend/**/"]
python_files = ["tests.py", "test_*.py", "*_test.py"]

[tool.coverage.run]
relative_files = true
omit = [
    "backend/**/tests.py",
]

[tool.uv.sources]
extended-dataset-profile = { git = "https://github.com/Beebucket/extended_dataset_profile", tag = "1.0.0" }
